requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
openai==1.3.0
python-dotenv==1.0.0
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0
watchdog==3.0.0
requests==2.31.0 